backend/app/retriever/faiss_retriever.py
"""FAISS retriever implementation"""
from typing import List, Dict, Any
import pickle
from pathlib import Path

from app.retriever.base import BaseRetriever
from app.config import settings
from app.utils.logger import get_logger

logger = get_logger(__name__)


class FAISSRetriever(BaseRetriever):
    """
    FAISS-based document retriever
    
    Inference-only: loads pre-built index from disk
    """
    
    def __init__(self):
        """Initialize FAISS retriever"""
        self._index = None
        self._documents = None
        self._embeddings = None
        logger.info("FAISSRetriever initialized (lazy loading)")
    
    def _load_index(self):
        """Lazy load FAISS index and documents"""
        if self._index is not None:
            return
        
        try:
            import faiss
            
            index_path = settings.VECTORSTORE_PATH / "faiss_index.bin"
            docs_path = settings.VECTORSTORE_PATH / "documents.pkl"
            
            if not index_path.exists():
                logger.warning(f"FAISS index not found at: {index_path}")
                return
            
            # Load FAISS index
            logger.info(f"Loading FAISS index from: {index_path}")
            self._index = faiss.read_index(str(index_path))
            
            # Load documents metadata
            if docs_path.exists():
                with open(docs_path, 'rb') as f:
                    self._documents = pickle.load(f)
                logger.info(f"Loaded {len(self._documents)} documents")
            
            logger.info("✅ FAISS index loaded successfully")
            
        except ImportError:
            logger.warning("faiss-cpu not installed. Install with: pip install faiss-cpu")
        except Exception as e:
            logger.error(f"Failed to load FAISS index: {str(e)}")
    
    def search(
        self,
        query: str,
        top_k: int = 5,
        score_threshold: float = 0.7
    ) -> List[Dict[str, Any]]:
        """
        Search for relevant documents using FAISS
        
        Args:
            query: Search query
            top_k: Number of results
            score_threshold: Minimum similarity score
            
        Returns:
            List of documents
        """
        self._load_index()
        
        if self._index is None or self._documents is None:
            logger.warning("FAISS index not loaded, returning empty results")
            return []
        
        try:
            # TODO: Implement actual embedding and search
            # For now, return placeholder
            logger.info(f"Searching for: {query[:50]}... (top_k={top_k})")
            
            # Placeholder: return first k documents
            results = []
            for i in range(min(top_k, len(self._documents))):
                results.append({
                    "content": self._documents[i].get("content", ""),
                    "metadata": self._documents[i].get("metadata", {}),
                    "score": 1.0 - (i * 0.1)  # Fake scores
                })
            
            return results
            
        except Exception as e:
            logger.error(f"Search error: {str(e)}")
            return []
    
    def get_stats(self) -> Dict[str, Any]:
        """Get FAISS index statistics"""
        self._load_index()
        
        return {
            "type": "FAISS",
            "index_loaded": self._index is not None,
            "num_documents": len(self._documents) if self._documents else 0,
            "index_size": self._index.ntotal if self._index else 0
        }
