backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# LLM Runtime
llama-cpp-python==0.3.16

# Vector Store
faiss-cpu==1.7.4
numpy==1.26.3

# Utils
python-dotenv==1.0.0

# Optional: for better logging
coloredlogs==15.0.1
